requirements.txt
opencv-python
numpy
matplotlib
scipy
Pillow
